refactor: ageout -> root/ageout.c
